requirements.txt
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.3
gevent==22.10.2
greenlet==2.0.2
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.2
packaging==23.1
pandas==2.0.0
Pillow==9.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tzdata==2023.3
websocket-client==1.5.1
zope.event==4.6
zope.interface==6.0
